completed vendor product handling part

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -8,6 +8,7 @@ use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class VendorProductImageGalleryController extends Controller
 {
@@ -19,6 +20,11 @@ class VendorProductImageGalleryController extends Controller
     public function index(Request $request, VendorProductImageGalleryDataTable $datatable)
     {
         $product = Product::findOrFail($request->product);
+
+        /** Check Product Vendor */
+        if($product->vendor_id != Auth::user()->vendor->id){
+            abort(404);
+        }
         return $datatable->render('vendor.product.image-gallery.index', compact('product'));
     }
 
@@ -83,6 +89,12 @@ class VendorProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $productImage = ProductImageGallery::findOrFail($id);
+
+        /** Check Product Vendor */
+        if($productImage->product->vendor_id != Auth::user()->vendor->id){
+            abort(404);
+        }
+        
         $this->deleteImage($productImage->image);
         $productImage->delete();
